+++ src/main/java/com/icomfortableworld/jwt/config/WebSecurityConfig.java
@@ -7,16 +7,15 @@ import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
 import org.springframework.security.config.http.SessionCreationPolicy;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 
 import com.icomfortableworld.jwt.JwtProvider;
 import com.icomfortableworld.jwt.security.JwtAuthenticationFilter;
-import com.icomfortableworld.jwt.security.JwtAuthorizationFilter;
-import com.icomfortableworld.jwt.security.UserDetailsServiceImpl;
+import com.icomfortableworld.jwt.security.MemberDetailsServiceImpl;
 
 import lombok.RequiredArgsConstructor;
 
@@ -26,8 +25,7 @@ import lombok.RequiredArgsConstructor;
 public class WebSecurityConfig {
 
 	private final JwtProvider jwtProvider;
-	private final UserDetailsServiceImpl userDetailsService;
-	private final AuthenticationConfiguration authenticationConfiguration;
+	private final MemberDetailsServiceImpl userDetailsService;
 
 	@Bean
 	public PasswordEncoder passwordEncoder() {
@@ -40,20 +38,13 @@ public class WebSecurityConfig {
 	}
 
 	@Bean
-	public JwtAuthenticationFilter jwtAuthenticationFilter() throws Exception {
-		JwtAuthenticationFilter filter = new JwtAuthenticationFilter(jwtProvider);
-		filter.setAuthenticationManager(authenticationManager(authenticationConfiguration));
-		return filter;
-	}
-
-	@Bean
-	public JwtAuthorizationFilter jwtAuthorizationFilter() {
-		return new JwtAuthorizationFilter(jwtProvider, userDetailsService);
+	public JwtAuthenticationFilter jwtAuthorizationFilter() {
+		return new JwtAuthenticationFilter(jwtProvider, userDetailsService);
 	}
 
 	@Bean
 	public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
-		http.csrf((csrf) -> csrf.disable());
+		http.csrf(AbstractHttpConfigurer::disable);
 
 		http.sessionManagement((sessionManagement) ->
 			sessionManagement.sessionCreationPolicy(SessionCreationPolicy.STATELESS)
@@ -66,11 +57,6 @@ public class WebSecurityConfig {
 				.requestMatchers("/api/version-1/members/**").permitAll()
 				.anyRequest().authenticated()
 		);
-
-		// 필터 관리
-		http.addFilterBefore(jwtAuthorizationFilter(), JwtAuthenticationFilter.class);
-		http.addFilterBefore(jwtAuthenticationFilter(), UsernamePasswordAuthenticationFilter.class);
-
 		return http.build();
 	}
 }
