--- src/main/java/com/icomfortableworld/jwt/security/JwtAuthorizationFilter.java
@@ -1,71 +1,0 @@
-package com.icomfortableworld.jwt.security;
-
-import java.io.IOException;
-
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContext;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.util.StringUtils;
-import org.springframework.web.filter.OncePerRequestFilter;
-
-import com.icomfortableworld.jwt.JwtProvider;
-
-import io.jsonwebtoken.Claims;
-import jakarta.servlet.FilterChain;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.extern.slf4j.Slf4j;
-
-@Slf4j(topic = "JWT 검증 및 인가")
-public class JwtAuthorizationFilter extends OncePerRequestFilter {
-
-	private final JwtProvider jwtProvider;
-	private final UserDetailsServiceImpl userDetailsService;
-
-	public JwtAuthorizationFilter(JwtProvider jwtProvider, UserDetailsServiceImpl userDetailsService) {
-		this.jwtProvider = jwtProvider;
-		this.userDetailsService = userDetailsService;
-	}
-
-	@Override
-	protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response,
-		FilterChain filterChain) throws ServletException, IOException {
-
-		String tokenValue = jwtProvider.getJwtFromHeader(request);
-
-		if (StringUtils.hasText(tokenValue)) {
-
-			if (!jwtProvider.validateToken(tokenValue)) {
-				log.error("Token Error");
-				return;
-			}
-
-			Claims info = jwtProvider.getUserInfoFromToken(tokenValue);
-
-			try {
-				setAuthentication(info.getSubject());
-			} catch (Exception e) {
-				log.error(e.getMessage());
-				return;
-			}
-		}
-
-		filterChain.doFilter(request, response);
-	}
-
-	public void setAuthentication(String username) {
-		SecurityContext context = SecurityContextHolder.createEmptyContext();
-		Authentication authentication = createAuthentication(username);
-		context.setAuthentication(authentication);
-
-		SecurityContextHolder.setContext(context);
-	}
-
-	private Authentication createAuthentication(String username) {
-		UserDetails userDetails = userDetailsService.loadUserByUsername(username);
-		return new UsernamePasswordAuthenticationToken(userDetails, null, userDetails.getAuthorities());
-	}
-}
